Slot suggestion algorithm v2

+++ app/src/main/java/com/sama/slotsuggestion/configuration/HeatMapConfiguration.kt
@@ -1,6 +1,6 @@
 package com.sama.slotsuggestion.configuration
 
-import com.sama.slotsuggestion.domain.WeightContext
+import com.sama.slotsuggestion.domain.v1.WeightContext
 import org.springframework.boot.context.properties.ConfigurationProperties
 import org.springframework.boot.context.properties.ConstructorBinding
 import org.springframework.context.annotation.Bean
@@ -9,7 +9,7 @@ import org.springframework.context.annotation.Configuration
 @ConstructorBinding
 @ConfigurationProperties(prefix = "sama.slotsuggestion.heat-map")
 class HeatMapConfiguration(
-    val intervalMinutes: Int,
+    val intervalMinutes: Long,
     val historicalDays: Long,
     val futureDays: Long,
 )
@@ -19,6 +19,6 @@ class WeightConfiguration {
 
     @Bean
     fun weightContext(heatMapConfiguration: HeatMapConfiguration): WeightContext {
-        return WeightContext(heatMapConfiguration.intervalMinutes, heatMapConfiguration.futureDays.toInt())
+        return WeightContext(heatMapConfiguration.intervalMinutes.toInt(), heatMapConfiguration.futureDays.toInt())
     }
 }
